Extract review bot wait and profile helpers
